pyproject.toml
[project]
name = "mcp-knowledge-base"
version = "0.1.0"
description = "Example MCP server to create a knowledge-base"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
 "mcp>=1.0.0",
 "python-dotenv>=1.0.1",
 "requests>=2.32.3",
]
[[project.authors]]
name = "Markus Pfundstein"
email = "markus@life-electronic.nl"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
mcp-knowledge-base = "mcp_knowledge_base:main"
